LayoutTests/http/tests/site-isolation/autofill-credentials-in-cross-origin-iframe.html

<!DOCTYPE html><!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script src="/js-test-resources/ui-helper.js"></script>
</head>
<body>
<script>
description("Test that inserting an autofill login credential fills the username and password inputs of a login form inside a cross-origin iframe when site isolation is enabled.");
jsTestIsAsync = true;

window.addEventListener("message", async (event) => {
    if (event.data.type === "credentials") {
        window.username = event.data.username;
        window.password = event.data.password;
        shouldBeEqualToString("username", "frederik");
        shouldBeEqualToString("password", "famos");
        finishJSTest();
    }
});

async function runTest() {
    const iframe = document.getElementById("frame");
    const iframeRect = iframe.getBoundingClientRect();

    const inputCenter = await new Promise(resolve => {
        window.addEventListener("message", (event) => {
            if (event.data.type === "inputRect")
                resolve(event.data);
        }, { once: true });
        iframe.contentWindow.postMessage({ type: "getInputRect" }, "*");
    });

    const clickX = iframeRect.left + inputCenter.x;
    const clickY = iframeRect.top + inputCenter.y;

    await UIHelper.activateAndWaitForInputSessionAt(clickX, clickY);

    await UIHelper.insertAutofillSuggestion("frederik", "famos");

    iframe.contentWindow.postMessage({ type: "getCredentials" }, "*");
}
</script>
<iframe id="frame" width="300" height="100" src="http://localhost:8000/site-isolation/resources/autofill-credentials-iframe.html"
onload="runTest()"></iframe>
</body>
</html>
