LayoutTests/http/tests/site-isolation/autofill-credentials-in-cross-origin-iframe-expected.txt

Test that inserting an autofill login credential fills the username and password inputs of a login form inside a cross-origin iframe when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS username is "frederik"
PASS password is "famos"
PASS successfullyParsed is true

TEST COMPLETE
